requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.6
Django==4.2.2
djangorestframework==3.14.0
dulwich==0.21.5
filelock==3.12.0
flake8==6.0.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pendulum==2.1.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.5.1
poetry-plugin-export==1.4.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
pywin32-ctypes==0.2.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
trove-classifiers==2023.5.24
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.0
webencodings==0.5.1
zipp==3.15.0
